Refactoring for simpler structure
