+++ .gitignore
@@ -1,0 +1,2 @@
+/env
+test_temper.py
